Document real backend, drop Spotify, embed cold seeds on demand
- Add ARCHITECTURE.md with Mermaid diagrams of every flow
- Rewrite CLAUDE.md to match the implementation (track_id ids, blended
  tags, MMR re-ranking, reject steering, playlists, Deezer/iTunes covers)
- Remove vestigial Spotify config: SPOTIFY_* env vars, render.yaml
  entries, and the stale spotify_id-era app/migrations/init.sql
- /recommendations now embeds cold seeds on demand, 404s on unknown
  track_ids, and guards all-zero vectors instead of returning [] for all
  three; widen embed_and_store_track listener_cap to float for inf cap

+++ app/services/ingest.py
@@ -3,7 +3,7 @@ from app.services import lastfm, embeddings
 from app.services.covers import get_cover_url
 
 
-def embed_and_store_track(artist: str, name: str, listener_cap: int) -> dict | None:
+def embed_and_store_track(artist: str, name: str, listener_cap: float) -> dict | None:
     """
     Ensure a track is embedded and stored in the songs table, fetching tags from
     Last.fm when we haven't seen it before. Returns the full song row
